tools: add a new import tool
This tool can be used to update the manifest-db from a gitlab ci
pipeline. CI pipelines are building all the possible manifest with the
Schutzfile's version of the CI. Thus, after a new run, we can use this
script to update the database.

Usage;

Firstly, run ci_import at the root of the manifest-db project:

./tools/ci_import --pipeline-id $pipelineid --token your-gitlab-token

Finally you can commit the modifications made to the test cases.