db: workflow to update the database
First a github action pulls the new manifests from composer. Creates a
commit out of the modification and pushes these to the gitlab repository
on a specific branch. Then it triggers a pipeline using the trigger
token.

Then the CI runs in "trigger" mode. In this mode it will compile all the
new manifests and produce the corresponding image-info in the artifacts.

Once this is done, in the final phase the CI will download all the
produced image-info and ammend the update db commit with them. Then it
will push that on github and open a PR.