tools/import: update manifest cleans image-info
We face an issue of inconsistency when we upgrade the manifests. Let's
say that, in the previous state of the repo we were running along a set
of distros that are reaching end of life support. These distros will no
longer be executed on the CI. Therefore, updating their manifests from
composer will lead to a unregenerated image-info. Having this
inconsistency in the database is wrong because, if osbuild then fails its
tests we can't know if it is because of a dirty entry or a failure in
the code base.

This commit put the image-info side of the db to an empty dict upon
update of the manifest side. Then when it'll be regenerated on CI, only
the image-info that were produced will be include in the database.