diff: filter tmpfiles_d if necessary
Some of the files scanned by tmpfiles_d may need to be filter out
because their content change for each build.
The design decision made is to store in the DataBase the list of files
a given distribution needs to filter out.

Filterdiff.py gets introduced and contains the necessary bits to perform
smart diffing between two image info. It possesses also a main function
and can be used to diff two files. Each file can either contain a
raw image-info json or be a DB entry.