osbuild-image-test: don't fail if there are no tests
Hard fail when there are no test manifests for the given distro and arch
make it unnecessarily complicated to enable new distro and generate
manifests for them (it would have to be done in two steps). There is
IMHO really no good reason to hard fail if there are no test manifests
for the selected distro and arch. Print a warning and exit gracefully.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>